requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
pydantic
